Reorganizing packages and removing unused code
