+++ lib/src/commonMain/kotlin/co/touchlab/kjwt/builder/JwtBuilder.kt
@@ -1,7 +1,5 @@
 package co.touchlab.kjwt.builder
 
-import co.touchlab.kjwt.cryptography.JweProcessor
-import co.touchlab.kjwt.cryptography.JwsProcessor
 import co.touchlab.kjwt.cryptography.SimpleKey
 import co.touchlab.kjwt.internal.encodeBase64Url
 import co.touchlab.kjwt.internal.encodeToBase64Url
@@ -11,12 +9,14 @@ import co.touchlab.kjwt.model.JwtPayload
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
-import co.touchlab.kjwt.model.crypto.CryptographyKotlinEncryptionProcessor
-import co.touchlab.kjwt.model.crypto.CryptographyKotlinIntegrityProcessor
 import co.touchlab.kjwt.model.registry.EncryptionKey
 import co.touchlab.kjwt.model.registry.JwtKeyRegistry
 import co.touchlab.kjwt.model.registry.SigningKey
 import co.touchlab.kjwt.model.registry.SigningKey.Identifier
+import co.touchlab.kjwt.processor.JweProcessor
+import co.touchlab.kjwt.processor.JwsProcessor
+import co.touchlab.kjwt.processor.crypto.CryptographyKotlinEncryptionProcessor
+import co.touchlab.kjwt.processor.crypto.CryptographyKotlinIntegrityProcessor
 import dev.whyoleg.cryptography.materials.key.Key
 import kotlinx.serialization.SerializationStrategy
 import kotlinx.serialization.json.Json
